+++ testRadar/testx.cpp
@@ -81,8 +81,8 @@ static TestStation* builder(const char* const fileName)
 //-----------------------------------------------------------------------------
 static void updateDataCB(int)
 {
-   double dt0 = 1.0/double(bgRate);
-   unsigned int millis = (unsigned int) (dt0 * 1000);
+   double dt0 = 1.0/static_cast<double>(bgRate);
+   unsigned int millis = static_cast<unsigned int>(dt0 * 1000);
    glutTimerFunc(millis, updateDataCB, 1);
 
    // Current time
@@ -134,8 +134,8 @@ int main(int argc, char* argv[])
 // Set timer for the background tasks
 // ---
 
-   double dt = 1.0/double(bgRate);
-   unsigned int millis = (unsigned int) (dt * 1000);
+   double dt = 1.0/static_cast<double>(bgRate);
+   unsigned int millis = static_cast<unsigned int>(dt * 1000);
 
    // ensure everything is reset
    station->updateData(dt);
